requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# NLP and ML libraries
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
spacy==3.7.2
scikit-learn==1.3.2

# Vector search and embeddings
faiss-cpu==1.7.4

# LangChain for text splitting and document processing
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52

# PDF processing
pypdf==3.17.0

# Utilities
numpy==1.24.3
requests==2.31.0

